# nvgpu reshape optimization: pass args instead of pointer to array