5,791us -> 2,968us

before:
```
Benchmarking ../ngraph-models/models/tensorflow/inference/resnet50/tf_function_ngraph_cluster_rn50_bs1.json, GPU backend, 10 iterations.
compile time: 7,191ms
14.3ms per iteration

---- Aggregate times per op type ----
Convolution     7,734us
Reshape         5,791us
BatchNorm         504us
Relu              325us
Add               205us
MaxPool            59us
Dot                34us
Sum                13us
Pad                11us
Divide              5us
Broadcast           5us
Result              1us
```
after:
```
Benchmarking ../ngraph-models/models/tensorflow/inference/resnet50/tf_function_ngraph_cluster_rn50_bs1.json, GPU backend, 10 iterations.
compile time: 7,043ms
11.5ms per iteration

---- Aggregate times per op type ----
Convolution     7,766us
Reshape         2,968us
BatchNorm         497us
Relu              334us
Add               202us
MaxPool            59us
Dot                34us
Sum                13us
Pad                10us
Divide              5us
Broadcast           5us
Result              1us
```